requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.1.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.27
protobuf==4.21.12
Werkzeug==1.0.1
zipp==3.17.0
python-dotenv
